rearranged values in the state
